# node/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Ethereum REST API
  
host: localhost:10010

basePath: /

schemes:
  - http
  - https
  
consumes:
  - application/json

produces:
  - application/json
  
paths:

###############################
########### MISC ##############
###############################

  /healthcheck:
    x-swagger-router-controller: misc.controller
    get:
      description: Test if web3 is connected
      operationId: healthcheck
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/genericObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"

###############################
########### SMART CONTRACT ####
###############################

  /setgreet:
    x-swagger-router-controller: smartContract.controller
    post:
      description: Modify greet
      operationId: setGreet
      parameters:
        - in: body
          name: newGreet
          description: New greet
          schema:
            $ref: "#/definitions/newGreet"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/genericObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
            
  /getgreet:
    x-swagger-router-controller: smartContract.controller
    get:
      description: Set new greet
      operationId: getGreet
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/genericObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
            

###############################
########## DEFINITIONS ########
###############################          
            
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:

###############################
########### ARGUMENTS #########
############################### 
        
  newGreet:
    properties:
      newgreet:
        type: string
    required:
      - newgreet

###############################
########## RESPONSES ##########
###############################  

  genericObjectResponse:
    type: object
      
###############################
############ ERRORS ###########
###############################        
    
  errorResponse:
    required:
      - message
    properties:
      message:
        type: string